// src/types/model.d.ts
export type salesDataType = {
  Category: string, 
  City: string,
  Country: string,
  "Customer ID": string,
  "Customer Name": string,
  Discount: string,
  "Order Date": string,
  "Order ID": string,
  "Postal Code": string,
  "Product ID": string,
  "Product Name": string,
  Profit: string,
  Quantity: string,
  Region: string,
  "Row ID": string,
  Sales: string,
  Segment: string,
  "Ship Date": string,
  "Ship Mode": string,
  State: string,
  "Sub-Category": string
  }; 

 